bug report
A bug was reported in which a management pack was not loading correctly.

The cause was that the ParentFolder (root folder) was not found, as it
did not include the .Console part.

I removed that and added the ! to ensure it was still unique in the
parent folder name, and its now working as expected.